… src/daft-logical-plan/src/optimization/rules/mod.rs
… src/daft-logical-plan/src/optimization/rules/push_down_anti_semi_join.rs