## Changes Made

Added an optimizer rule `PushDownAntiSemiJoin`, which pushes anti and semi joins through inner joins, projects, filters, sorts, and distincts.

I decided to do this instead of adding anti and semi joins to our brute force join reorderer because it is simpler and will largely produce the same results, especially with TPC-H. 

More specifically, join graph creation makes some assumptions about the uniqueness of input column names that I'm not sure will hold for anti and semi joins, since their columns are not preserved in the output. I was not confident in my ability to modify the join reordering algorithm to properly fix this within a reasonable time. We will want to implement a more efficient reordering algorithm in the future anyway, so I do not want to spend a significant effort on patching the current one up.

### TPC-H Q18 
#### Results
| Benchmark | Before | After | Improvement |
|--------|--------|--------|--------|
| Local (SF 100) | 56.79s | 10.38s | 5.47x |
| Distributed (SF 1000) | 285.4s | 66.12s | 4.32x |

#### Optimized Plan (Before)
```mermaid
flowchart TD
subgraph optimized["Optimized LogicalPlan"]
direction TB
optimizedLimit0["Limit: 100
Stats = { Approx num rows = 100, Approx size bytes = 1.56 KiB, Accumulated
selectivity = 0.00 }"]
optimizedSort1["Sort: Sort by = (col(o_totalprice), descending, nulls first), (col(o_orderdate),
ascending, nulls last)
Stats = { Approx num rows = 5,665,935, Approx size bytes = 86.46 MiB,
Accumulated selectivity = 0.12 }"]
optimizedAggregate2["Aggregation: sum(col(l_quantity))
Group by = col(c_name), col(c_custkey), col(o_orderkey), col(o_orderdate),
col(o_totalprice)
Output schema = c_name#Utf8, c_custkey#Int64, o_orderkey#Int64,
o_orderdate#Date, o_totalprice#Float64, l_quantity#Int64
Stats = { Approx num rows = 5,665,935, Approx size bytes = 86.46 MiB,
Accumulated selectivity = 0.12 }"]
optimizedProject3["Project: col(l_quantity), col(c_name), col(c_custkey), col(o_orderkey),
col(o_orderdate), col(o_totalprice)
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedJoin4["Join: Type = Semi
Strategy = Auto
On = col(left.o_orderkey#Int64) == col(right.l_orderkey#Int64)
Output schema = o_orderkey#Int64, c_custkey#Int64, c_name#Utf8,
o_totalprice#Float64, o_orderdate#Date, l_quantity#Int64
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedProject5["Project: col(o_orderkey), col(c_custkey), col(c_name), col(o_totalprice),
col(o_orderdate), col(l_quantity)
Stats = { Approx num rows = 46,594,852, Approx size bytes = 722.09 MiB,
Accumulated selectivity = 0.95 }"]
optimizedJoin6["Join: Type = Inner
Strategy = Auto
On = col(left.o_orderkey#Int64) == col(right.l_orderkey#Int64)
Output schema = c_custkey#Int64, c_name#Utf8, o_orderkey#Int64,
o_totalprice#Float64, o_orderdate#Date, l_orderkey#Int64, l_quantity#Int64
Stats = { Approx num rows = 46,594,852, Approx size bytes = 722.09 MiB,
Accumulated selectivity = 0.95 }"]
optimizedProject7["Project: col(c_custkey), col(c_name), col(o_orderkey), col(o_totalprice),
col(o_orderdate)
Stats = { Approx num rows = 18,273,257, Approx size bytes = 496.66 MiB,
Accumulated selectivity = 0.95 }"]
optimizedJoin8["Join: Type = Inner
Strategy = Auto
On = col(left.c_custkey#Int64) == col(right.o_custkey#Int64)
Output schema = c_custkey#Int64, c_name#Utf8, o_orderkey#Int64, o_custkey#Int64,
o_totalprice#Float64, o_orderdate#Date
Stats = { Approx num rows = 18,273,257, Approx size bytes = 496.66 MiB,
Accumulated selectivity = 0.95 }"]
optimizedProject9["Project: col(C_CUSTKEY) as c_custkey, col(C_NAME) as c_name
Stats = { Approx num rows = 3,308,528, Approx size bytes = 89.14 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource10["Num Scan Tasks = 32
File schema = C_CUSTKEY#Int64, C_NAME#Utf8, C_ADDRESS#Utf8, C_NATIONKEY#Int64,
C_PHONE#Utf8, C_ACCTBAL#Float64, C_MKTSEGMENT#Utf8, C_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [C_CUSTKEY, C_NAME]
Output schema = C_CUSTKEY#Int64, C_NAME#Utf8
Stats = { Approx num rows = 3,308,528, Approx size bytes = 89.14 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource10 --> optimizedProject9
optimizedProject9 --> optimizedJoin8
optimizedProject11["Project: col(O_ORDERKEY) as o_orderkey, col(O_CUSTKEY) as o_custkey,
col(O_TOTALPRICE) as o_totalprice, col(O_ORDERDATE) as o_orderdate
Stats = { Approx num rows = 18,273,257, Approx size bytes = 496.66 MiB,
Accumulated selectivity = 0.95 }"]
optimizedSource12["Num Scan Tasks = 32
File schema = O_ORDERKEY#Int64, O_CUSTKEY#Int64, O_ORDERSTATUS#Utf8,
O_TOTALPRICE#Float64, O_ORDERDATE#Date, O_ORDERPRIORITY#Utf8, O_CLERK#Utf8,
O_SHIPPRIORITY#Int64, O_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [O_ORDERKEY, O_CUSTKEY, O_TOTALPRICE, O_ORDERDATE]
Filter pushdown = not(is_null(col(O_ORDERKEY)))
Output schema = O_ORDERKEY#Int64, O_CUSTKEY#Int64, O_TOTALPRICE#Float64,
O_ORDERDATE#Date
Stats = { Approx num rows = 18,273,257, Approx size bytes = 496.66 MiB,
Accumulated selectivity = 0.95 }"]
optimizedSource12 --> optimizedProject11
optimizedProject11 --> optimizedJoin8
optimizedJoin8 --> optimizedProject7
optimizedProject7 --> optimizedJoin6
optimizedProject13["Project: col(L_ORDERKEY) as l_orderkey, col(L_QUANTITY) as l_quantity
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource14["Num Scan Tasks = 32
File schema = L_ORDERKEY#Int64, L_PARTKEY#Int64, L_SUPPKEY#Int64,
L_LINENUMBER#Int64, L_QUANTITY#Int64, L_EXTENDEDPRICE#Float64,
L_DISCOUNT#Float64, L_TAX#Float64, L_RETURNFLAG#Utf8, L_LINESTATUS#Utf8,
L_SHIPDATE#Date, L_COMMITDATE#Date, L_RECEIPTDATE#Date, L_SHIPINSTRUCT#Utf8,
L_SHIPMODE#Utf8, L_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [L_ORDERKEY, L_QUANTITY]
Output schema = L_ORDERKEY#Int64, L_QUANTITY#Int64
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource14 --> optimizedProject13
optimizedProject13 --> optimizedJoin6
optimizedJoin6 --> optimizedProject5
optimizedProject5 --> optimizedJoin4
optimizedProject15["Project: col(l_orderkey)
Stats = { Approx num rows = 7,455,177, Approx size bytes = 113.76 MiB,
Accumulated selectivity = 0.15 }"]
optimizedFilter16["Filter: col((l_quantity.local_sum() > Literal(Int64(300))))
Stats = { Approx num rows = 7,455,177, Approx size bytes = 113.76 MiB,
Accumulated selectivity = 0.15 }"]
optimizedProject17["Project: col(l_orderkey), col(l_quantity.local_sum()) > lit(300) as
(l_quantity.local_sum() > Literal(Int64(300)))
Stats = { Approx num rows = 37,275,881, Approx size bytes = 568.78 MiB,
Accumulated selectivity = 0.76 }"]
optimizedFilter18["Filter: not(is_null(col(l_orderkey)))
Stats = { Approx num rows = 37,275,881, Approx size bytes = 568.78 MiB,
Accumulated selectivity = 0.76 }"]
optimizedAggregate19["Aggregation: sum(col(l_quantity)) as l_quantity.local_sum()
Group by = col(l_orderkey)
Output schema = l_orderkey#Int64, l_quantity.local_sum()#Int64
Stats = { Approx num rows = 39,237,769, Approx size bytes = 598.72 MiB,
Accumulated selectivity = 0.80 }"]
optimizedProject20["Project: col(L_QUANTITY) as l_quantity, col(L_ORDERKEY) as l_orderkey
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource21["Num Scan Tasks = 32
File schema = L_ORDERKEY#Int64, L_PARTKEY#Int64, L_SUPPKEY#Int64,
L_LINENUMBER#Int64, L_QUANTITY#Int64, L_EXTENDEDPRICE#Float64,
L_DISCOUNT#Float64, L_TAX#Float64, L_RETURNFLAG#Utf8, L_LINESTATUS#Utf8,
L_SHIPDATE#Date, L_COMMITDATE#Date, L_RECEIPTDATE#Date, L_SHIPINSTRUCT#Utf8,
L_SHIPMODE#Utf8, L_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [L_QUANTITY, L_ORDERKEY]
Output schema = L_ORDERKEY#Int64, L_QUANTITY#Int64
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource21 --> optimizedProject20
optimizedProject20 --> optimizedAggregate19
optimizedAggregate19 --> optimizedFilter18
optimizedFilter18 --> optimizedProject17
optimizedProject17 --> optimizedFilter16
optimizedFilter16 --> optimizedProject15
optimizedProject15 --> optimizedJoin4
optimizedJoin4 --> optimizedProject3
optimizedProject3 --> optimizedAggregate2
optimizedAggregate2 --> optimizedSort1
optimizedSort1 --> optimizedLimit0
end
```

#### Optimized Plan (After)

```mermaid
flowchart TD
subgraph optimized["Optimized LogicalPlan"]
direction TB
optimizedLimit0["Limit: 100
Stats = { Approx num rows = 100, Approx size bytes = 1.56 KiB, Accumulated
selectivity = 0.00 }"]
optimizedSort1["Sort: Sort by = (col(o_totalprice), descending, nulls first), (col(o_orderdate),
ascending, nulls last)
Stats = { Approx num rows = 5,665,935, Approx size bytes = 86.46 MiB,
Accumulated selectivity = 0.12 }"]
optimizedAggregate2["Aggregation: sum(col(l_quantity))
Group by = col(c_name), col(c_custkey), col(o_orderkey), col(o_orderdate),
col(o_totalprice)
Output schema = c_name#Utf8, c_custkey#Int64, o_orderkey#Int64,
o_orderdate#Date, o_totalprice#Float64, l_quantity#Int64
Stats = { Approx num rows = 5,665,935, Approx size bytes = 86.46 MiB,
Accumulated selectivity = 0.12 }"]
optimizedProject3["Project: col(l_quantity), col(c_name), col(c_custkey), col(o_orderkey),
col(o_orderdate), col(o_totalprice)
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedProject4["Project: col(c_custkey), col(c_name), col(o_orderkey), col(o_totalprice),
col(o_orderdate), col(l_orderkey), col(l_quantity)
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedJoin5["Join: Type = Inner
Strategy = Auto
On = col(left.c_custkey#Int64) == col(right.o_custkey#Int64)
Output schema = c_custkey#Int64, c_name#Utf8, o_orderkey#Int64, o_custkey#Int64,
o_totalprice#Float64, o_orderdate#Date, l_orderkey#Int64, l_quantity#Int64
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedProject6["Project: col(C_CUSTKEY) as c_custkey, col(C_NAME) as c_name
Stats = { Approx num rows = 3,308,528, Approx size bytes = 89.14 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource7["Num Scan Tasks = 32
File schema = C_CUSTKEY#Int64, C_NAME#Utf8, C_ADDRESS#Utf8, C_NATIONKEY#Int64,
C_PHONE#Utf8, C_ACCTBAL#Float64, C_MKTSEGMENT#Utf8, C_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [C_CUSTKEY, C_NAME]
Output schema = C_CUSTKEY#Int64, C_NAME#Utf8
Stats = { Approx num rows = 3,308,528, Approx size bytes = 89.14 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource7 --> optimizedProject6
optimizedProject6 --> optimizedJoin5
optimizedJoin8["Join: Type = Inner
Strategy = Auto
On = col(left.o_orderkey#Int64) == col(right.l_orderkey#Int64)
Output schema = o_orderkey#Int64, o_custkey#Int64, o_totalprice#Float64,
o_orderdate#Date, l_orderkey#Int64, l_quantity#Int64
Stats = { Approx num rows = 7,082,419, Approx size bytes = 109.76 MiB,
Accumulated selectivity = 0.14 }"]
optimizedProject9["Project: col(O_ORDERKEY) as o_orderkey, col(O_CUSTKEY) as o_custkey,
col(O_TOTALPRICE) as o_totalprice, col(O_ORDERDATE) as o_orderdate
Stats = { Approx num rows = 7,082,419, Approx size bytes = 108.07 MiB,
Accumulated selectivity = 0.14 }"]
optimizedJoin10["Join: Type = Semi
Strategy = Auto
On = col(left.O_ORDERKEY#Int64) == col(right.l_orderkey#Int64)
Output schema = O_ORDERKEY#Int64, O_CUSTKEY#Int64, O_TOTALPRICE#Float64,
O_ORDERDATE#Date
Stats = { Approx num rows = 7,082,419, Approx size bytes = 108.07 MiB,
Accumulated selectivity = 0.14 }"]
optimizedSource11["Num Scan Tasks = 32
File schema = O_ORDERKEY#Int64, O_CUSTKEY#Int64, O_ORDERSTATUS#Utf8,
O_TOTALPRICE#Float64, O_ORDERDATE#Date, O_ORDERPRIORITY#Utf8, O_CLERK#Utf8,
O_SHIPPRIORITY#Int64, O_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [O_ORDERKEY, O_CUSTKEY, O_TOTALPRICE, O_ORDERDATE]
Filter pushdown = not(is_null(col(O_ORDERKEY)))
Output schema = O_ORDERKEY#Int64, O_CUSTKEY#Int64, O_TOTALPRICE#Float64,
O_ORDERDATE#Date
Stats = { Approx num rows = 18,273,257, Approx size bytes = 496.66 MiB,
Accumulated selectivity = 0.95 }"]
optimizedSource11 --> optimizedJoin10
optimizedProject12["Project: col(l_orderkey)
Stats = { Approx num rows = 7,455,177, Approx size bytes = 113.76 MiB,
Accumulated selectivity = 0.15 }"]
optimizedFilter13["Filter: col((l_quantity.local_sum() > Literal(Int64(300))))
Stats = { Approx num rows = 7,455,177, Approx size bytes = 113.76 MiB,
Accumulated selectivity = 0.15 }"]
optimizedProject14["Project: col(l_orderkey), col(l_quantity.local_sum()) > lit(300) as
(l_quantity.local_sum() > Literal(Int64(300)))
Stats = { Approx num rows = 37,275,881, Approx size bytes = 568.78 MiB,
Accumulated selectivity = 0.76 }"]
optimizedFilter15["Filter: not(is_null(col(l_orderkey)))
Stats = { Approx num rows = 37,275,881, Approx size bytes = 568.78 MiB,
Accumulated selectivity = 0.76 }"]
optimizedAggregate16["Aggregation: sum(col(l_quantity)) as l_quantity.local_sum()
Group by = col(l_orderkey)
Output schema = l_orderkey#Int64, l_quantity.local_sum()#Int64
Stats = { Approx num rows = 39,237,769, Approx size bytes = 598.72 MiB,
Accumulated selectivity = 0.80 }"]
optimizedProject17["Project: col(L_QUANTITY) as l_quantity, col(L_ORDERKEY) as l_orderkey
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource18["Num Scan Tasks = 32
File schema = L_ORDERKEY#Int64, L_PARTKEY#Int64, L_SUPPKEY#Int64,
L_LINENUMBER#Int64, L_QUANTITY#Int64, L_EXTENDEDPRICE#Float64,
L_DISCOUNT#Float64, L_TAX#Float64, L_RETURNFLAG#Utf8, L_LINESTATUS#Utf8,
L_SHIPDATE#Date, L_COMMITDATE#Date, L_RECEIPTDATE#Date, L_SHIPINSTRUCT#Utf8,
L_SHIPMODE#Utf8, L_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [L_QUANTITY, L_ORDERKEY]
Output schema = L_ORDERKEY#Int64, L_QUANTITY#Int64
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource18 --> optimizedProject17
optimizedProject17 --> optimizedAggregate16
optimizedAggregate16 --> optimizedFilter15
optimizedFilter15 --> optimizedProject14
optimizedProject14 --> optimizedFilter13
optimizedFilter13 --> optimizedProject12
optimizedProject12 --> optimizedJoin10
optimizedJoin10 --> optimizedProject9
optimizedProject9 --> optimizedJoin8
optimizedProject19["Project: col(L_ORDERKEY) as l_orderkey, col(L_QUANTITY) as l_quantity
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource20["Num Scan Tasks = 32
File schema = L_ORDERKEY#Int64, L_PARTKEY#Int64, L_SUPPKEY#Int64,
L_LINENUMBER#Int64, L_QUANTITY#Int64, L_EXTENDEDPRICE#Float64,
L_DISCOUNT#Float64, L_TAX#Float64, L_RETURNFLAG#Utf8, L_LINESTATUS#Utf8,
L_SHIPDATE#Date, L_COMMITDATE#Date, L_RECEIPTDATE#Date, L_SHIPINSTRUCT#Utf8,
L_SHIPMODE#Utf8, L_COMMENT#Utf8
Partitioning keys = []
Projection pushdown = [L_ORDERKEY, L_QUANTITY]
Output schema = L_ORDERKEY#Int64, L_QUANTITY#Int64
Stats = { Approx num rows = 49,047,212, Approx size bytes = 760.10 MiB,
Accumulated selectivity = 1.00 }"]
optimizedSource20 --> optimizedProject19
optimizedProject19 --> optimizedJoin8
optimizedJoin8 --> optimizedJoin5
optimizedJoin5 --> optimizedProject4
optimizedProject4 --> optimizedProject3
optimizedProject3 --> optimizedAggregate2
optimizedAggregate2 --> optimizedSort1
optimizedSort1 --> optimizedLimit0
end
```
## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130 for docs review)